+++ core/osal/src/ut-stubs/utstub-helpers.h
@@ -93,7 +93,7 @@ typedef enum
 typedef struct
 {
     uint32 LastIssueNumber;
-    uint8 ValidBits[OSAL_MAX_VALID_PER_TYPE];
+    uint32 ValidBits[OSAL_MAX_VALID_PER_TYPE];
 } UT_ObjTypeState_t;
